requirements.txt
flask
flask_cors
pandas
pillow
requests
pickle5
roboflow
scikit-learn